SpaceTime S3 Raster Catalog support
